refactor: Restructure AdventyPay component and styles, update import paths, and remove obsolete files for improved organization and maintainability
